Move the bot class to the main file
